restaurant owner portal

+++ src/utils/axiosInstance.ts
@@ -1,5 +1,6 @@
 import axios, { type AxiosResponse } from "axios";
 import { PUBLIC_ROUTE } from "../enums";
+import toast from "react-hot-toast";
 
 const axiosInstance = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
@@ -11,9 +12,12 @@ const axiosInstance = axios.create({
 axiosInstance.interceptors.request.use(
   (config) => {
     if (!config.url?.includes("auth/roles")) {
-      const token = JSON.parse(localStorage.getItem("token") || "");
-      if (!config.url?.includes("roles") && token && config.headers) {
-        config.headers = config.headers || {};
+      const localToken = JSON.parse(localStorage.getItem("token") || "null");
+      const urlParams = new URLSearchParams(window.location.search);
+      const sharedToken = urlParams.get("token");
+      const token = sharedToken || localToken;
+
+      if (token && config.headers) {
         config.headers["token"] = token;
       }
     }
@@ -25,6 +29,10 @@ axiosInstance.interceptors.request.use(
 axiosInstance.interceptors.response.use(
   (response: AxiosResponse) => response,
   async (error: any) => {
+    if (error.code === "ERR_NETWORK") {
+      toast.error("Something wrong happened!");
+      return;
+    }
     if (error.response?.data?.message === "jwt expired") {
       localStorage.clear();
       window.location.href = PUBLIC_ROUTE.SIGNIN;
